requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
pandas==2.2.2
SQLAlchemy==2.0.32
Pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
ftfy==6.2.3
regex==2024.5.15
tqdm==4.66.5
git+https://github.com/openai/CLIP.git
